// material-kit-react/src/store/types.ts
export enum DatasetExploreActionType {
    GET_CLASSES = 'GET_CLASSES',
    GET_CLASSES_SUCCESS = 'GET_CLASSES_SUCCESS',
    GET_CLASSES_ERROR = 'GET_CLASSES_ERROR',
    GET_HOLE_DATASET = 'GET_HOLE_DATASET',
    GET_HOLE_DATASET_SUCCESS = 'GET_HOLE_DATASET_SUCCESS',
    GET_HOLE_DATASET_ERROR = 'GET_HOLE_DATASET_ERROR',
    GET_SINGLE_DATASET = 'GET_SINGLE_DATASET',
    GET_SINGLE_DATASET_SUCCESS = 'GET_SINGLE_DATASET_SUCCESS',
    GET_SINGLE_DATASET_ERROR = 'GET_SINGLE_DATASET_ERROR',

    POST_FAVORATE_DATASET = 'POST_FAVORATE_DATASET',
    POST_FAVORATE_DATASET_SUCCESS = 'POST_FAVORATE_DATASET_SUCCESS',
    POST_FAVORATE_DATASET_ERROR = 'POST_FAVORATE_DATASET_ERROR',


    
}